+++ app/scripts/lib/lforms-data.js
@@ -3059,10 +3059,11 @@
 
 
     /**
-     * Compare if the two given codings are equal. Details see the code below.
-     * @param coding1
-     * @param coding2
-     * @return {boolean} true if the two codings are considerer equal, false otherwise.
+     * Compare if the two given codings are equal. A "coding" is a hash that may have any or all of the
+     * following three fields: code, system, and text. See the code below for the precise matching rules.
+     * @param coding1 the first coding object
+     * @param coding2 the second coding object
+     * @return {boolean} true if the two codings are considered equal, false otherwise.
      * @private
      */
     _codingsEqual: function(coding1, coding2) {
@@ -3077,6 +3078,7 @@
       }
       return false;
     },
+
 
     /**
      * Check if a source item's value meet a skip logic condition/trigger